Test for sid collision #36
A bit contrived, but does demonstrate the problem. If two processes
randomly generate the same sid, a collision is possible. There is a
window between the `get` call in `sid_collision?` and the `set` call in
`set_session` where another process could set the same sid.

This should be fixed by using `setnx` in `sid_collision?`.