Add :ttl configuration option
**Why**: The :expire_after config option controls both the Redis
session expiration and the expiration of the Rails session cookie.
Adding a new config option :ttl allows those controls to be
managed separately.

The most common case would be if you wanted the Rails cookie
to expire when the browser is closed, as it would if :expire_after
were set to `nil`.